move magic_link and security question to views
